made the markdown box a control of its own, that can be included in otehr components
