+++ apps/ds-ui/src/styles/_read.scss
@@ -24,13 +24,6 @@
   }
 }
 
-.notes-preview {
-  @extend .bg-light;
-  @extend .mx-1;
-  @extend .my-2;
-  @extend .border;
-}
-
 .passage-card {
   @extend .h-100;
   @extend .shadow;
