+++ apps/ds-ui/src/styles/ds-styles.scss
@@ -159,3 +159,4 @@ footer {
 @import './home';
 @import './pray';
 @import './plans';
+@import './markdown';
